+++ tests-integration/tests/hll_test/main.rs
@@ -16,6 +16,6 @@
 // under the License.
 
 mod bounds;
-mod display;
+mod summary;
 mod union;
 mod update;